+++ src/main/java/Items/Weapons/Sword.java
@@ -2,7 +2,7 @@ package Items.Weapons;
 
 public class Sword extends Weapon {
 
-    public Sword(String name, double attackPower) {
+    public Sword(String name, int attackPower) {
         super(name, attackPower);
     }
 }
