+++ src/main/java/Players/Healer/Cleric.java
@@ -4,7 +4,7 @@ import Players.Player;
 
 public class Cleric extends Healer {
 
-    public Cleric(String name, double healthPoints) {
+    public Cleric(String name, int healthPoints) {
         super(name, healthPoints);
     }
