+++ src/main/java/Players/MeleeFighter/Knight.java
@@ -1,7 +1,7 @@
 package Players.MeleeFighter;
 
 public class Knight extends MeleeFighter {
-    public Knight(String name, double healthPoints) {
+    public Knight(String name, int healthPoints) {
         super(name, healthPoints);
     }
 }
